requirements.txt
agora-token-builder==1.0.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
idna==3.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
django-crispy-forms==1.14.0
djangorestframework==3.14.0
django-rest-framework-braces==0.3.4